feat(json-concerns): extract product serialization into ProductsJson concern
